# web/assets/openapi.yaml
openapi: 3.0.3
info:
  title: RenoP HTTP API
  version: "2.0"
  contact:
    url: https://github.com/404Setup/RenoP
  description: |
    Management and registry APIs for RenoP, a self-hosted artifact repository server
    supporting Maven, unstructured file storage, npm, Cargo (Rust sparse index), and Docker/OCI registries.

    Default listen address: `0.0.0.0:3000`.

    **Route layout:**
    - Management API: `/api/...`
    - Maven artifact storage: `/{repo}/{path}` (Maven 2 layout)
    - Unstructured file storage: `/{repo}/{path}`
    - Docker/OCI Distribution API: `/v2/...`
    - Cargo sparse registry: `/{repo}/config.json` and `/{repo}/api/v1/...`
    - npm registry: `/{repo}/{package}` and `/{repo}/-/...`
    - Javadoc preview: `/javadoc/{repo}/{path}`
    - Cargo-doc preview: `/cargodoc/{repo}/{path}`

    **Wire format:**
    Most management endpoints use `application/x-protobuf`. Schemas here describe the
    logical fields (snake_case) for documentation and client stub generation.
    See `proto/api/v1/api.proto` for canonical protobuf definitions.
    Docker/OCI, npm, and Cargo endpoints use JSON.

    **Error bodies:**
    Management errors are plain text (`Unauthorized`, `Forbidden`, `Not found`).
    Docker browser-management errors additionally expose a stable `X-Renop-Error-Code` header for localization.
    Docker Distribution API errors follow `{"errors":[{"code":"...","message":"...","detail":"..."}]}`.
...

  license:
    name: MPL-2.0
...
servers:
...
components:
  securitySchemes:
    sessionCookie:
      type: apiKey
      in: cookie
      name: renop_session
      description: HttpOnly session cookie issued by `POST /api/auth/login`. Idle timeout is approximately 7 days.
    basicAuth:
      type: http
      scheme: basic
      description: |
        `username:password` or `username:API-token` encoded as HTTP Basic credentials.
        Basic credentials are accepted only by standard package protocols, Docker token exchange, and chunked upload.
    bearerAuth:
      type: http
      scheme: bearer
      description: |
        A prefixed API-token secret. API endpoints intersect the token's scopes with the owning account's current
        permissions. Query-string credentials and `Authorization: Session` are rejected.
...
  schemas:
...
    ReviewTask:
      type: object
      required: [ id, kind, resource_type, resource_key, resource_name, review_team_prefix, requested_by, status, created_at ]
      properties:
        id: { type: string, format: uuid }
        kind: { type: string, enum: [ super_team_transfer, publication ] }
        resource_type:
          type: string
          enum: [ docker_image, npm_package, cargo_package, maven_artifact, maven_domain ]
        repository:
          type: string
          description: Empty for process-wide Maven publishing domains.
        resource_key:
          type: string
          maxLength: 1024
        resource_name:
          type: string
          maxLength: 512
        resource_version: { type: string, maxLength: 255 }
        source_team_prefix: { type: string, maxLength: 64 }
        target_team_prefix:
          type: string
          maxLength: 64
          description: Target owner for a transfer, or the team that approved a T2 package-creation request.
        review_team_prefix:
          type: string
          maxLength: 64
          description: Non-empty while a team task is waiting for one of its T3/T4 managers; cleared when a package-creation task advances to repository review.
        requested_by: { type: string }
        status: { type: string, enum: [ pending, approved, rejected, cancelled ] }
        decision_reason:
          type: string
          maxLength: 512
          description: Preset publication reasons use `preset:<code>`; custom reasons use `custom:<text>`; transfers retain their reason text.
        decided_by:
          type: string
          description: Present only for system administrators; requester and moderator responses omit reviewer identity.
        created_at: { type: integer, format: int64 }
        updated_at: { type: integer, format: int64 }
        decided_at: { type: integer, format: int64 }
        file_count: { type: integer, minimum: 0, maximum: 256 }
        total_size: { type: integer, format: int64, minimum: 0 }
    ReviewFile:
      type: object
      required: [ id, path, name, size, critical, added_at ]
      properties:
        id: { type: string, pattern: "^[a-f0-9]{64}$" }
        path: { type: string, maxLength: 1024 }
        name: { type: string }
        size: { type: integer, format: int64, minimum: 0 }
        critical: { type: boolean }
        added_at: { type: integer, format: int64 }
    ReviewTaskPage:
      type: object
      required: [ tasks, total, limit, offset, view ]
      properties:
        tasks:
          type: array
          items: { $ref: "#/components/schemas/ReviewTask" }
        total: { type: integer, minimum: 0 }
        limit: { type: integer, minimum: 1, maximum: 100 }
        offset: { type: integer, minimum: 0 }
        view: { type: string, enum: [ reviewer, requested ] }
    SuperTeamTransferReviewRequest:
      type: object
      required: [ resource_type, resource_key, target_team_prefix ]
      properties:
        resource_type:
          type: string
          enum: [ docker_image, npm_package, cargo_package, maven_artifact, maven_domain ]
        repository:
          type: string
          description: Required for package and image resources; omitted for Maven publishing domains.
        resource_key:
          type: string
          description: Canonical package name, image name, domain, or `groupId:artifactId` coordinate.
        target_team_prefix:
          type: string
          maxLength: 64
          description: Target team for transfer-in; empty for a return to personal ownership.
...
paths:
...
  /api/reviews:
    get:
      tags: [ reviews ]
      summary: List review tasks
      description: >
        Returns a bounded reviewer or requester page. Team transfers and T2 package-creation requests are visible to
        current T3/T4 team managers. Publication reviews are visible to repository-scoped moderators and system
        administrators only after any required team approval. This route accepts a browser session only; Basic
        credentials and Bearer API tokens are denied.
      operationId: listReviewTasks
      security:
        - sessionCookie: [ ]
      parameters:
        - name: view
          in: query
          schema: { type: string, enum: [ reviewer, requested ], default: reviewer }
        - name: status
          in: query
          schema: { type: string, enum: [ pending, approved, rejected, cancelled, all ], default: pending }
        - name: types
          in: query
          description: Optional comma-separated resource-type filter.
          style: form
          explode: false
          schema:
            type: array
            maxItems: 5
            items:
              type: string
              enum: [ docker_image, npm_package, cargo_package, maven_artifact, maven_domain ]
        - { name: limit, in: query, schema: { type: integer, minimum: 1, maximum: 100, default: 20 } }
        - { name: offset, in: query, schema: { type: integer, minimum: 0, default: 0 } }
      responses:
        "200":
          description: Bounded review page.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ReviewTaskPage" }
        "400": { description: Invalid view, status, type, or page boundary. }
        "401": { description: A browser session is required. }
  /api/reviews/super-team-transfers:
    post:
      tags: [ reviews ]
      summary: Request a global-team ownership transfer
      description: >
        The caller must own the project or domain at L4, or hold current repository/system administration authority.
        Transfer-in also requires membership in the target team.
        A T3 or T4 manager of the reviewing team or a system administrator must decide the request. Direct team-to-team transfers,
        mirrored resources, and returns of namespaced Docker or scoped npm resources to personal ownership are rejected.
      operationId: createSuperTeamTransferReview
      security:
        - sessionCookie: [ ]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/SuperTeamTransferReviewRequest" }
      responses:
        "201":
          description: Pending review created.
          headers:
            Location:
              schema: { type: string }
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ReviewTask" }
        "400": { description: Invalid resource identity or namespace/team mismatch. }
        "403": { description: The caller does not own or administer the resource, or is not a member of the target team. }
        "409": { description: This resource already has a pending transfer, ownership changed, or transfer-out is restricted. }
  /api/reviews/{id}/decision:
    post:
      tags: [ reviews ]
      summary: Decide one pending review
      description: >
        A T3/T4 manager decides team transfers and T2 package-creation requests. If repository creation review is also
        enabled, team approval keeps the same task pending, clears `review_team_prefix`, and forwards it to repository
        moderators; it does not reserve the name. A repository moderator or system administrator decides that second
        stage after the five-second settling window. Publication rejection uses a preset reason or bounded custom
        reason. Compare-and-set transitions prevent concurrent decisions from applying twice.
      operationId: decideReviewTask
      security:
        - sessionCookie: [ ]
      parameters:
        - { name: id, in: path, required: true, schema: { type: string, format: uuid } }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [ decision ]
              properties:
                decision: { type: string, enum: [ approved, rejected ] }
                reason:
                  type: string
                  maxLength: 505
                  description: Required for a custom publication rejection or any transfer rejection; ignored for approval.
                reason_code:
                  type: string
                  enum: [ invalid_metadata, quality, policy_violation, copyright, malware, custom ]
                  description: Required for publication rejection. `custom` also requires `reason`.
      responses:
        "200":
          description: Review decided, or a team-approved package creation advanced to its still-pending repository-review stage.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ReviewTask" }
        "400": { description: Invalid decision or missing rejection reason. }
        "403": { description: The caller is not an eligible independent reviewer. }
        "404": { description: Review task not found. }
        "409": { description: Another reviewer decided the task, the resource changed, or files are still arriving. }
  /api/reviews/{id}/files:
    get:
      tags: [ reviews ]
      summary: List publication review files
      description: Returns the bounded repository-relative file list to the requester, assigned repository moderator, or system administrator.
      operationId: listReviewFiles
      security:
        - sessionCookie: [ ]
      parameters:
        - { name: id, in: path, required: true, schema: { type: string, format: uuid } }
      responses:
        "200":
          description: Review file list.
          content:
            application/json:
              schema:
                type: object
                required: [ task_id, files, total ]
                properties:
                  task_id: { type: string, format: uuid }
                  files:
                    type: array
                    maxItems: 256
                    items: { $ref: "#/components/schemas/ReviewFile" }
                  total: { type: integer, minimum: 0, maximum: 256 }
        "403": { description: The caller cannot inspect this publication. }
        "404": { description: Review task not found. }
  /api/reviews/{id}/files/{file_id}:
    get:
      tags: [ reviews ]
      summary: Download one hidden publication review file
      operationId: downloadReviewFile
      security:
        - sessionCookie: [ ]
      parameters:
        - { name: id, in: path, required: true, schema: { type: string, format: uuid } }
        - { name: file_id, in: path, required: true, schema: { type: string, pattern: "^[a-f0-9]{64}$" } }
      responses:
        "200":
          description: Hidden review file stream.
          content:
            application/octet-stream:
              schema: { type: string, format: binary }
        "403": { description: The caller cannot inspect this publication. }
        "404": { description: Review task or file not found. }
  /api/reviews/{id}:
    delete:
      tags: [ reviews ]
      summary: Cancel a pending review request
      description: Only the original requester may cancel a pending ownership-transfer task. Publication reviews cannot be cancelled through this route.
      operationId: cancelReviewTask
      security:
        - sessionCookie: [ ]
      parameters:
        - { name: id, in: path, required: true, schema: { type: string, format: uuid } }
      responses:
        "200":
          description: Cancelled review record.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ReviewTask" }
        "403": { description: The caller did not request this task. }
        "404": { description: Review task not found. }
        "409": { description: The task is no longer pending. }

  # ========== Maven domain registry ==========
...